+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 1.3.14
+
+### Patch Changes
+
+- Updated dependencies [[`02872400`](https://github.com/LekoArts/gatsby-themes/commit/0287240022c308a7d1fcc8af348ee7d21bca0dd5)]:
+  - @lekoarts/gatsby-theme-minimal-blog@6.0.0
+
+## 1.3.13
+
+### Patch Changes
+
+- Updated dependencies [[`ee969f3`](https://github.com/LekoArts/gatsby-themes/commit/ee969f30037fa99232292014431854773735d0a0)]:
+  - @lekoarts/gatsby-theme-minimal-blog@5.0.0
+
 ## 1.3.12
 
 ### Patch Changes